Refactor tools module into package
