+++ twin_generator/pipeline_helpers.py
@@ -5,7 +5,9 @@ from typing import Any
 
 from agents.run import Runner as AgentsRunner  # type: ignore
 
-from .tools import calc_answer_tool, make_html_table_tool, render_graph_tool
+from .tools.calc import calc_answer_tool
+from .tools.graph import render_graph_tool
+from .tools.html_table import make_html_table_tool
 from .utils import get_final_output, safe_json
 
 __all__ = [
